Fix push to unqualified destination
When pushing a revision to a reference that doesn't exist in the remote
git cannot know what we want to do. In that case we check if the tag
exists locally, then we assume we want to push it as tag too, if not
we assume we want is as branch.